requirements.txt
boto3==1.10.18
botocore==1.13.18
certifi==2019.9.11
cffi==1.13.2
chardet==3.0.4
Click==7.0
cryptography==2.8
docutils==0.15.2
Flask==1.1.1
future==0.18.2
gunicorn==20.0.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
jmespath==0.9.4
joblib==0.14.0
MarkupSafe==1.1.1
numpy==1.17.4
pycparser==2.19
python-dateutil==2.8.0
pytorch-ignite
python-telegram-bot==12.2.0
pytorch-transformers==1.2.0
regex==2019.11.1
requests==2.7.0
s3transfer==0.2.1
sacremoses==0.0.35
sentencepiece==0.1.83
six==1.13.0
tensorboardX==1.8
tensorflow  # for tensorboardX
torch==1.3.1
tornado==6.0.3
tqdm==4.38.0
urllib3==1.25.7
Werkzeug==0.16.0
